added the initial conversations that happen in level one. They will need polishing
