+++ app/phaser/LevelOne/Conversations/Ghost.ts
@@ -1,0 +1,122 @@
+import Phaser from "phaser";
+
+interface DialogueChoice {
+  text: string;
+  next: number; // index of next dialogue node
+}
+
+interface DialogueNode {
+  text: string;
+  choices?: DialogueChoice[];
+}
+
+export default class Ghost extends Phaser.Scene {
+  private dialogueNodes: DialogueNode[] = [
+    {
+      text: "Blood Ghost: YOU TOOK IT FROM ME! MY HUSK! ILL KILL YOU ILL KILL YOU ILL KILL YOU... *His punches phase right through you*",
+    },
+  ];
+
+  private currentNodeIndex: number = 0;
+  private dialogueText!: Phaser.GameObjects.Text;
+  private choiceTexts: Phaser.GameObjects.Text[] = [];
+  private music!: Phaser.Sound.BaseSound;
+
+  constructor() {
+    super({ key: "Ghost" });
+  }
+
+  preload() {
+    this.load.image("ghostOneConvo", "/assets/conversations/ghost.png");
+    this.load.audio("ghostOneMusic", "/assets/music/mystery.mp3");
+  }
+
+  create() {
+    // Background portrait
+    const portrait = this.add
+      .image(this.scale.width / 2, this.scale.height / 2, "ghostOneConvo")
+      .setOrigin(0.5);
+    portrait.displayWidth = this.scale.width;
+    portrait.displayHeight = this.scale.height;
+
+    // Dialogue box
+    this.add.rectangle(
+      this.scale.width / 2,
+      this.scale.height - 150,
+      this.scale.width,
+      290,
+      0x000000,
+      0.4
+    );
+
+    this.dialogueText = this.add.text(150, this.scale.height - 270, "", {
+      fontSize: "26px",
+      color: "#ffffff",
+      wordWrap: { width: this.scale.width - 300 },
+    });
+
+    this.music = this.sound.add("ghostOneMusic", { loop: true, volume: 1 });
+    this.music.play();
+
+    // Show first node
+    this.showNode(0);
+
+    // Input: pick choices with number keys
+    this.input.keyboard!.on("keydown", (event: KeyboardEvent) => {
+      const key = parseInt(event.key);
+      if (!isNaN(key)) {
+        const choice =
+          this.dialogueNodes[this.currentNodeIndex].choices?.[key - 1];
+        if (choice) {
+          this.showNode(choice.next);
+        }
+      }
+    });
+  }
+
+  private showNode(index: number) {
+    this.currentNodeIndex = index;
+    const node = this.dialogueNodes[index];
+
+    this.dialogueText.setText(node.text);
+
+    // Remove old choices
+    this.choiceTexts.forEach((c) => c.destroy());
+    this.choiceTexts = [];
+
+    // If no choices, check if end
+    if (!node.choices || node.choices.length === 0) {
+      this.add.text(
+        180,
+        this.scale.height - 110,
+        "Press space to exit conversation",
+        {
+          fontSize: "24px",
+          color: "#ffcc00",
+          wordWrap: { width: this.scale.width - 300 },
+        }
+      );
+      this.input.keyboard!.once("keydown-SPACE", () => {
+        this.music.stop();
+        this.scene.stop();
+        this.scene.resume("SceneOne");
+      });
+      return;
+    }
+
+    // Show new choices
+    node.choices.forEach((choice, i) => {
+      const choiceText = this.add.text(
+        180,
+        this.scale.height - 110 + i * 40,
+        choice.text,
+        {
+          fontSize: "24px",
+          color: "#ffcc00",
+          wordWrap: { width: this.scale.width - 300 },
+        }
+      );
+      this.choiceTexts.push(choiceText);
+    });
+  }
+}
